refactor(message-handler): modularize message handling logic
- 將訊息處理邏輯重構為多個專用處理器，提升代碼的可讀性與維護性。
- 新增 blob-handler.js、right-click-handler.js 和 element-registration-handler.js，分別處理 Blob URL 註冊、右鍵點擊和元素註冊的訊息。
- 更新 message-handler.js 以根據訊息類型路由到相應的處理器，簡化主邏輯結構。
- 此變更有助於清晰地分離不同功能的邏輯，為未來的擴展提供更好的基礎。
